+++ packages/typescript-estree/src/convert.ts
@@ -10,7 +10,7 @@ import type {
 import type { SemanticOrSyntacticError } from './semantic-or-syntactic-errors';
 import type { TSESTree, TSESTreeToTSNode, TSNode } from './ts-estree';
 
-import { checkModifiers } from './check-modifiers';
+import { checkSyntaxError } from './check-syntax-errors';
 import { getDecorators, getModifiers } from './getModifiers';
 import {
   canContainDirective,
@@ -105,58 +105,12 @@ export class Converter {
     this.options = { ...options };
   }
 
-  #checkForStatementDeclaration(
-    initializer: ts.ForInitializer,
-    kind: ts.SyntaxKind.ForInStatement | ts.SyntaxKind.ForOfStatement,
-  ): void {
-    const loop =
-      kind === ts.SyntaxKind.ForInStatement ? 'for...in' : 'for...of';
-    if (ts.isVariableDeclarationList(initializer)) {
-      if (initializer.declarations.length !== 1) {
-        this.#throwError(
-          initializer,
-          `Only a single variable declaration is allowed in a '${loop}' statement.`,
-        );
-      }
-      const declaration = initializer.declarations[0];
-      if (declaration.initializer) {
-        this.#throwError(
-          declaration,
-          `The variable declaration of a '${loop}' statement cannot have an initializer.`,
-        );
-      } else if (declaration.type) {
-        this.#throwError(
-          declaration,
-          `The variable declaration of a '${loop}' statement cannot have a type annotation.`,
-        );
-      }
-      if (
-        kind === ts.SyntaxKind.ForInStatement &&
-        initializer.flags & ts.NodeFlags.Using
-      ) {
-        this.#throwError(
-          initializer,
-          "The left-hand side of a 'for...in' statement cannot be a 'using' declaration.",
-        );
-      }
-    } else if (
-      !isValidAssignmentTarget(initializer) &&
-      initializer.kind !== ts.SyntaxKind.ObjectLiteralExpression &&
-      initializer.kind !== ts.SyntaxKind.ArrayLiteralExpression
-    ) {
-      this.#throwError(
-        initializer,
-        `The left-hand side of a '${loop}' statement must be a variable or a property access.`,
-      );
-    }
-  }
-
-  #checkModifiers(node: ts.Node): void {
+  #checkSyntaxError(node: ts.Node): void {
     if (this.options.allowInvalidAST) {
       return;
     }
 
-    checkModifiers(node);
+    checkSyntaxError(node);
   }
 
   #throwError(
@@ -528,7 +482,7 @@ export class Converter {
       return null;
     }
 
-    this.#checkModifiers(node);
+    this.#checkSyntaxError(node);
 
     const pattern = this.allowPattern;
     if (allowPattern != null) {
@@ -896,7 +850,6 @@ export class Converter {
         });
 
       case SyntaxKind.ForInStatement:
-        this.#checkForStatementDeclaration(node.initializer, node.kind);
         return this.createNode<TSESTree.ForInStatement>(node, {
           type: AST_NODE_TYPES.ForInStatement,
           body: this.convertChild(node.statement),
@@ -905,13 +858,9 @@ export class Converter {
         });
 
       case SyntaxKind.ForOfStatement: {
-        this.#checkForStatementDeclaration(node.initializer, node.kind);
         return this.createNode<TSESTree.ForOfStatement>(node, {
           type: AST_NODE_TYPES.ForOfStatement,
-          await: Boolean(
-            node.awaitModifier &&
-            node.awaitModifier.kind === SyntaxKind.AwaitKeyword,
-          ),
+          await: node.awaitModifier?.kind === SyntaxKind.AwaitKeyword,
           body: this.convertChild(node.statement),
           left: this.convertPattern(node.initializer),
           right: this.convertChild(node.expression),
@@ -974,9 +923,9 @@ export class Converter {
       }
 
       case SyntaxKind.VariableDeclaration: {
-        const definite = !!node.exclamationToken;
+        const hasExclamationToken = !!node.exclamationToken;
 
-        if (definite) {
+        if (hasExclamationToken) {
           if (node.initializer) {
             this.#throwError(
               node,
@@ -990,6 +939,90 @@ export class Converter {
           }
         }
 
+        if (node.parent.kind === SyntaxKind.VariableDeclarationList) {
+          const variableDeclarationList = node.parent;
+          const kind = getDeclarationKind(variableDeclarationList);
+
+          if (
+            (variableDeclarationList.parent.kind ===
+              SyntaxKind.ForInStatement ||
+              variableDeclarationList.parent.kind ===
+                SyntaxKind.ForStatement) &&
+            (kind === 'using' || kind === 'await using')
+          ) {
+            if (!node.initializer) {
+              this.#throwError(
+                node,
+                `'${kind}' declarations may not be initialized in for statement.`,
+              );
+            }
+
+            if (node.name.kind !== SyntaxKind.Identifier) {
+              this.#throwError(
+                node.name,
+                `'${kind}' declarations may not have binding patterns.`,
+              );
+            }
+          }
+
+          if (
+            variableDeclarationList.parent.kind === SyntaxKind.VariableStatement
+          ) {
+            const variableStatement = variableDeclarationList.parent;
+
+            if (kind === 'using' || kind === 'await using') {
+              if (!node.initializer) {
+                this.#throwError(
+                  node,
+                  `'${kind}' declarations must be initialized.`,
+                );
+              }
+              if (node.name.kind !== SyntaxKind.Identifier) {
+                this.#throwError(
+                  node.name,
+                  `'${kind}' declarations may not have binding patterns.`,
+                );
+              }
+            }
+
+            const hasDeclareKeyword = hasModifier(
+              SyntaxKind.DeclareKeyword,
+              variableStatement,
+            );
+
+            // Definite assignment only allowed for non-declare let and var
+            if (
+              (hasDeclareKeyword ||
+                ['await using', 'const', 'using'].includes(kind)) &&
+              hasExclamationToken
+            ) {
+              this.#throwError(
+                node,
+                `A definite assignment assertion '!' is not permitted in this context.`,
+              );
+            }
+
+            if (
+              hasDeclareKeyword &&
+              node.initializer &&
+              (['let', 'var'].includes(kind) || node.type)
+            ) {
+              this.#throwError(
+                node,
+                `Initializers are not permitted in ambient contexts.`,
+              );
+            }
+            // Theoretically, only certain initializers are allowed for declare const,
+            // (TS1254: A 'const' initializer in an ambient context must be a string
+            // or numeric literal or literal enum reference.) but we just allow
+            // all expressions
+
+            // Note! No-declare does not mean the variable is not ambient, because
+            // it can be further nested in other declare contexts. Therefore we cannot
+            // check for const initializers.
+          }
+        }
+
         const init = this.convertChild(node.initializer);
         const id = this.convertBindingNameWithTypeAnnotation(
           node.name,
@@ -998,77 +1031,28 @@ export class Converter {
         );
         return this.createNode<TSESTree.VariableDeclarator>(node, {
           type: AST_NODE_TYPES.VariableDeclarator,
-          definite,
+          definite: hasExclamationToken,
           id,
           init,
         });
       }
 
       case SyntaxKind.VariableStatement: {
-        const result = this.createNode<TSESTree.VariableDeclaration>(node, {
-          type: AST_NODE_TYPES.VariableDeclaration,
-          declarations: this.convertChildren(node.declarationList.declarations),
-          declare: hasModifier(SyntaxKind.DeclareKeyword, node),
-          kind: getDeclarationKind(node.declarationList),
-        });
+        const declarations = node.declarationList.declarations;
 
-        if (!result.declarations.length) {
+        if (!declarations.length) {
           this.#throwError(
             node,
             'A variable declaration list must have at least one variable declarator.',
           );
         }
-        if (result.kind === 'using' || result.kind === 'await using') {
-          node.declarationList.declarations.forEach((declaration, i) => {
-            if (result.declarations[i].init == null) {
-              this.#throwError(
-                declaration,
-                `'${result.kind}' declarations must be initialized.`,
-              );
-            }
-            if (result.declarations[i].id.type !== AST_NODE_TYPES.Identifier) {
-              this.#throwError(
-                declaration.name,
-                `'${result.kind}' declarations may not have binding patterns.`,
-              );
-            }
-          });
-        }
-        // Definite assignment only allowed for non-declare let and var
-        if (
-          result.declare ||
-          ['await using', 'const', 'using'].includes(result.kind)
-        ) {
-          node.declarationList.declarations.forEach((declaration, i) => {
-            if (result.declarations[i].definite) {
-              this.#throwError(
-                declaration,
-                `A definite assignment assertion '!' is not permitted in this context.`,
-              );
-            }
-          });
-        }
-        if (result.declare) {
-          node.declarationList.declarations.forEach((declaration, i) => {
-            if (
-              result.declarations[i].init &&
-              (['let', 'var'].includes(result.kind) ||
-                result.declarations[i].id.typeAnnotation)
-            ) {
-              this.#throwError(
-                declaration,
-                `Initializers are not permitted in ambient contexts.`,
-              );
-            }
-          });
-          // Theoretically, only certain initializers are allowed for declare const,
-          // (TS1254: A 'const' initializer in an ambient context must be a string
-          // or numeric literal or literal enum reference.) but we just allow
-          // all expressions
-        }
-        // Note! No-declare does not mean the variable is not ambient, because
-        // it can be further nested in other declare contexts. Therefore we cannot
-        // check for const initializers.
+
+        const result = this.createNode<TSESTree.VariableDeclaration>(node, {
+          type: AST_NODE_TYPES.VariableDeclaration,
+          declarations: this.convertChildren(declarations),
+          declare: hasModifier(SyntaxKind.DeclareKeyword, node),
+          kind: getDeclarationKind(node.declarationList),
+        });
 
         /**
          * Semantically, decorators are not allowed on variable declarations,
@@ -1082,30 +1066,12 @@ export class Converter {
 
       // mostly for for-of, for-in
       case SyntaxKind.VariableDeclarationList: {
-        const result = this.createNode<TSESTree.VariableDeclaration>(node, {
+        return this.createNode<TSESTree.VariableDeclaration>(node, {
           type: AST_NODE_TYPES.VariableDeclaration,
           declarations: this.convertChildren(node.declarations),
           declare: false,
           kind: getDeclarationKind(node),
         });
-
-        if (result.kind === 'using' || result.kind === 'await using') {
-          node.declarations.forEach((declaration, i) => {
-            if (result.declarations[i].init != null) {
-              this.#throwError(
-                declaration,
-                `'${result.kind}' declarations may not be initialized in for statement.`,
-              );
-            }
-            if (result.declarations[i].id.type !== AST_NODE_TYPES.Identifier) {
-              this.#throwError(
-                declaration.name,
-                `'${result.kind}' declarations may not have binding patterns.`,
-              );
-            }
-          });
-        }
-        return result;
       }
 
       // Expressions
@@ -1597,10 +1563,8 @@ export class Converter {
         } else {
           result = this.createNode<TSESTree.Property>(node, {
             type: AST_NODE_TYPES.Property,
-            computed: Boolean(
-              node.propertyName &&
-              node.propertyName.kind === SyntaxKind.ComputedPropertyName,
-            ),
+            computed:
+              node.propertyName?.kind === SyntaxKind.ComputedPropertyName,
             key: this.convertChild(node.propertyName ?? node.name),
             kind: 'init',
             method: false,